Correção e botão voltar

+++ db/migrate/20210917145457_add_fields_to_property.rb
@@ -2,6 +2,6 @@ class AddFieldsToProperty < ActiveRecord::Migration[6.1]
   def change
     add_column :properties, :bathrooms, :integer
     add_column :properties, :pets, :boolean
-    add_column :properties, :daily_rate, :integer
+    add_column :properties, :daily_rate, :decimal
   end
 end
